# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vidhai ERP API — multi-site mushroom farming production orchestration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ooty/rooms:
    get:
      operationId: listOotyRooms
      tags: [ooty]
      responses:
        "200":
          description: All rooms with current status
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OotyRoom" }
    post:
      operationId: createOotyRoom
      tags: [ooty]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OotyRoomInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyRoom" }
  /ooty/rooms/{id}:
    get:
      operationId: getOotyRoom
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Room detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyRoomDetail" }
    patch:
      operationId: updateOotyRoom
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OotyRoomUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyRoom" }
    delete:
      operationId: deleteOotyRoom
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "409":
          description: Room has active batch
  /ooty/growing-batches:
    get:
      operationId: listOotyGrowingBatches
      tags: [ooty]
      responses:
        "200":
          description: All growing batches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OotyGrowingBatch" }
    post:
      operationId: createOotyGrowingBatch
      tags: [ooty]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OotyGrowingBatchInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyGrowingBatch" }
  /ooty/growing-batches/{id}:
    get:
      operationId: getOotyGrowingBatch
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Growing batch detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyGrowingBatchDetail" }
    patch:
      operationId: updateOotyGrowingBatch
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OotyGrowingBatchUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyGrowingBatch" }
  /ooty/growing-batches/{id}/advance:
    post:
      operationId: advanceOotyPhase
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OotyPhaseAdvanceInput" }
      responses:
        "200":
          description: Phase advanced
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyGrowingBatch" }
  /ooty/growing-batches/{id}/observations:
    get:
      operationId: listOotyObservations
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Observations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OotyObservation" }
    post:
      operationId: addOotyObservation
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OotyObservationInput" }
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyObservation" }
  /ooty/growing-batches/{id}/harvests:
    get:
      operationId: listOotyHarvests
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Harvests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OotyHarvest" }
    post:
      operationId: addOotyHarvest
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OotyHarvestInput" }
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OotyHarvest" }
  /ooty/growing-batches/{id}/approve-phase:
    post:
      operationId: approveOotyPhase
      tags: [ooty]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PhaseApprovalInput" }
      responses:
        "200":
          description: Approval recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PhaseApproval" }

  # ── TRACEABILITY ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OotyRoom:
      type: object
      required: [id, name, locationId, status]
      properties:
        id: { type: integer }
        name: { type: string }
        locationId: { type: integer }
        status:
          type: string
          enum: [idle, active, maintenance]
        capacity: { type: ["integer", "null"] }
        currentGrowingBatchId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        currentBatch: { type: object, nullable: true }
        alertLevel: { type: string }
    OotyRoomDetail:
      type: object
      required: [id, name, locationId, status, batches]
      properties:
        id: { type: integer }
        name: { type: string }
        locationId: { type: integer }
        status: { type: string }
        capacity: { type: ["integer", "null"] }
        currentGrowingBatchId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        batches:
          type: array
          items: { $ref: "#/components/schemas/OotyGrowingBatch" }
    OotyRoomInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        capacity: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
    OotyRoomUpdate:
      type: object
      properties:
        name: { type: string }
        status: { type: string }
        capacity: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        currentGrowingBatchId: { type: ["integer", "null"] }
    OotyGrowingBatch:
      type: object
      required: [id, batchCode, roomId, currentPhase, status, createdAt]
      properties:
        id: { type: integer }
        batchCode: { type: string }
        roomId: { type: integer }
        annurBatchId: { type: ["integer", "null"] }
        coimBatchId: { type: ["integer", "null"] }
        currentPhase:
          type: string
          enum: [SPAWN_RUN, CASING_RUN, DF, COOKOUT, COMPLETED]
        phaseEnteredAt: { type: ["string", "null"] }
        status: { type: string }
        spawnRunStartDate: { type: ["string", "null"] }
        casingAppliedDate: { type: ["string", "null"] }
        cookoutDate: { type: ["string", "null"] }
        substrateWeightKg: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    OotyGrowingBatchDetail:
      type: object
      required: [id, batchCode, roomId, currentPhase, status, createdAt, dayInPhase, observations, harvests, approvals]
      properties:
        id: { type: integer }
        batchCode: { type: string }
        roomId: { type: integer }
        annurBatchId: { type: ["integer", "null"] }
        coimBatchId: { type: ["integer", "null"] }
        currentPhase: { type: string }
        phaseEnteredAt: { type: ["string", "null"] }
        dayInPhase: { type: ["integer", "null"] }
        status: { type: string }
        spawnRunStartDate: { type: ["string", "null"] }
        casingAppliedDate: { type: ["string", "null"] }
        cookoutDate: { type: ["string", "null"] }
        substrateWeightKg: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        observations:
          type: array
          items: { $ref: "#/components/schemas/OotyObservation" }
        harvests:
          type: array
          items: { $ref: "#/components/schemas/OotyHarvest" }
        approvals:
          type: array
          items: { $ref: "#/components/schemas/PhaseApproval" }
    OotyGrowingBatchInput:
      type: object
      required: [roomId]
      properties:
        roomId: { type: integer }
        annurBatchId: { type: ["integer", "null"] }
        coimBatchId: { type: ["integer", "null"] }
        spawnRunStartDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    OotyGrowingBatchUpdate:
      type: object
      properties:
        status: { type: string }
        notes: { type: ["string", "null"] }
        casingAppliedDate: { type: ["string", "null"] }
        cookoutDate: { type: ["string", "null"] }
        substrateWeightKg: { type: ["number", "null"] }
    OotyPhaseAdvanceInput:
      type: object
      required: [nextPhase]
      properties:
        nextPhase:
          type: string
          enum: [CASING_RUN, DF, COOKOUT, COMPLETED]
        notes: { type: ["string", "null"] }
        cookoutDate: { type: ["string", "null"] }
        substrateWeightKg: { type: ["number", "null"] }
    OotyObservation:
      type: object
      required: [id, growingBatchId, observationDate, observationType, createdAt]
      properties:
        id: { type: integer }
        growingBatchId: { type: integer }
        observationDate: { type: string }
        temperatureCelsius: { type: ["number", "null"] }
        observationNote: { type: ["string", "null"] }
        observationType: { type: string }
        createdAt: { type: string }
    OotyObservationInput:
      type: object
      required: [observationDate]
      properties:
        observationDate: { type: string }
        temperatureCelsius: { type: ["number", "null"] }
        observationNote: { type: ["string", "null"] }
        observationType: { type: string }
    OotyHarvest:
      type: object
      required: [id, growingBatchId, harvestDate, weightKg, flushNumber, createdAt]
      properties:
        id: { type: integer }
        growingBatchId: { type: integer }
        harvestDate: { type: string }
        weightKg: { type: number }
        mushroomCount: { type: ["integer", "null"] }
        avgWeightG: { type: ["number", "null"] }
        qualityNote: { type: ["string", "null"] }
        flushNumber: { type: integer }
        createdAt: { type: string }
    OotyHarvestInput:
      type: object
      required: [harvestDate, weightKg]
      properties:
        harvestDate: { type: string }
        weightKg: { type: number }
        mushroomCount: { type: ["integer", "null"] }
        avgWeightG: { type: ["number", "null"] }
        qualityNote: { type: ["string", "null"] }
        flushNumber: { type: integer }
    PhaseApproval:
      type: object
      required: [id, entityType, entityId, fromPhase, toPhase, createdAt]
      properties:
        id: { type: integer }
        entityType: { type: string }
        entityId: { type: integer }
        fromPhase: { type: string }
        toPhase: { type: string }
        decision: { type: ["string", "null"] }
        approvedByUserId: { type: ["integer", "null"] }
        decidedAt: { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        createdAt: { type: string }
    PhaseApprovalInput:
      type: object
      required: [fromPhase, toPhase, decision]
      properties:
        fromPhase: { type: string }
        toPhase: { type: string }
        decision:
          type: string
          enum: [approve, reject]
        rejectionReason: { type: ["string", "null"] }

    # ── TRACEABILITY ─────────────────────────────────────────────────────────────
...